Move gdpr and helauth views and urls to the api package
