+++ tilavarauspalvelu/urls.py
@@ -7,25 +7,15 @@ from graphene_file_upload.django import FileUploadGraphQLView
 
 from api.legacy_rest_api.urls import legacy_outer
 from api.webhooks.urls import webhook_router
-from users.views import TilavarauspalveluGDPRAPIView, TVPLoginView, TVPLogoutCompleteView, TVPLogoutView
-
-helusers_pattern = (
-    [
-        path("login/", TVPLoginView.as_view(), name="auth_login"),
-        path("logout/", TVPLogoutView.as_view(), name="auth_logout"),
-        path("logout/complete/", TVPLogoutCompleteView.as_view(), name="auth_logout_complete"),
-    ],
-    "helusers",
-)
 
 urlpatterns = [
     path("graphql/", csrf_exempt(FileUploadGraphQLView.as_view(graphiql=settings.DEBUG))),  # NOSONAR
     path("admin/", admin.site.urls),
     path("v1/", include(legacy_outer.urls)),
     path("v1/webhook/", include(webhook_router.urls)),
-    path("gdpr/v1/user/<str:uuid>/", TilavarauspalveluGDPRAPIView.as_view(), name="gdpr_v1"),
     path("pysocial/", include("social_django.urls", namespace="social")),
-    path("helauth/", include(helusers_pattern)),
+    path("helauth/", include("api.helauth.urls")),
+    path("gdpr/v1/", include("api.gdpr.urls")),
     path("tinymce/", include("tinymce.urls")),
     path("", include("django_prometheus.urls")),
 ]
